FE-1552: Move single-consumer modules beside their consumers
The compute activity list sits under the summary that renders it, the range-draw translator under the batch instantiator, pair selection inside the shader generator, and the bin-value summary under the plot hook that reads it. The session stops re-exporting the cell helpers only its own modules use, and the translator's random-draw oracle test becomes a fixed two-run fixture.
