+++ libs/@hashintel/petrinaut-core/src/webgpu/compile-net-shader/pair-selection.ts
